Lua 5.3 compatibility
The upcoming release of TeX Live 2018 will switch from Lua 5.2 to Lua 5.3 (at least pretest does).  This new release introduces a separation between integer and floating point types (a.k.a. "number").  Value of type "number" can no longer be formatted with the "%d" format string and have to be converted to integers first.  While Lua 5.3 introduces integer division via the double-slash operator "a // b", it is better to use "math.floor(a / b)" to maintain backwards compatibility with Lua 5.2.